backend/requirements.txt
# ── Existing packages (legacy server.py) ──────────────────────────────────────
fastapi==0.110.1
uvicorn[standard]==0.25.0
motor==3.3.1
pymongo==4.5.0
bcrypt==4.1.3
PyJWT==2.12.1
python-dotenv==1.2.2
pydantic[email]==2.12.5
python-multipart==0.0.22
starlette==0.37.2
anyio==4.13.0
requests==2.32.5
openpyxl==3.1.5

# ── Production backend (backend/app/) ─────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.1
pydantic-settings>=2.2.1
firebase-admin==6.5.0
boto3>=1.34.0
pillow>=10.3.0
apscheduler>=3.10.0
